+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "BusinessMonitor",
-  "version": "0.1.0+25",
+  "version": "0.1.0+26",
   "private": true,
   "type": "module",
   "scripts": {
